+++ triplea/service/repository/export/unified_export_json/__init__.py
@@ -5,6 +5,8 @@ from triplea.service.repository.export.unified_export_json.arxiv import (
 from triplea.service.repository.export.unified_export_json.pubmed import (
     _json_converter_01_pubmed,
 )
+from triplea.service.repository.export.unified_export_json.scopus import _json_converter_01_scopus
+from triplea.service.repository.export.unified_export_json.wos import _json_converter_01_wos
 
 
 def json_converter_01(article: Article):
@@ -16,6 +18,10 @@ def json_converter_01(article: Article):
         json_article = _json_converter_01_pubmed(article)
     elif article.SourceBank == SourceBankType.ARXIV:
         json_article = _json_converter_01_arxiv(article)
+    elif article.SourceBank == SourceBankType.WOS:
+        json_article = _json_converter_01_wos(article)
+    elif article.SourceBank == SourceBankType.SCOPUS:
+        json_article = _json_converter_01_scopus(article)
     else:
         raise NotImplementedError